Fixes for report component of preflight
- move into separate file
- fix typing for report and main
